services/telemetry-ingest/requirements.txt
grpcio>=1.59.0
grpcio-tools>=1.59.0
grpcio-health-checking>=1.59.0
redis[hiredis]>=5.0.1
aiokafka>=0.10.0
pydantic>=2.6.4
pydantic-settings>=2.2.0
python-dotenv>=1.0.1
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-otlp>=1.21.0
prometheus-client>=0.19.0
structlog>=23.0.0
tenacity>=8.2.0





